feature_extraction/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
numpy==1.24.1
pandas==1.5.2
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
typing-extensions==4.4.0
urllib3==1.26.13
